Update default address settings and query adjustments
The default user Id in the create-address.dto file is now set to 1. Also, adjustments have been made to the query handling in todos.service.ts and users.service.ts. In todos.service.ts, the 'addRelationsAndJoin' function was replaced with manual 'leftJoinAndSelect' calls, while in users.service.ts, 'addRelationsAndJoin' is used to join the user